Fix TypeError in creation of QRect
... expecting ints but getting floats

+++ python/stack_of_tasks/ui/traits_mapping/painter_entry/data_painter.py
@@ -57,7 +57,7 @@ class MatrixPainter(Delegate_Painter):
         rect.setSize(content_size)
 
         if (m := shape[0]) > 0:
-            dm = rect.height() / m
+            dm = int(rect.height() / m)
         else:
             dm = 0
 
@@ -73,7 +73,7 @@ class MatrixPainter(Delegate_Painter):
 
         if len(shape) > 1:
             n = shape[1]
-            dn = rect.width() / n
+            dn = int(rect.width() / n)
 
             lns.extend(
                 chain.from_iterable(
